+++ pippy/utils.py
@@ -3,7 +3,6 @@ import logging
 from typing import Dict
 
 import torch
-import torch.distributed as dist
 from torch import fx
 from torch.export.unflatten import InterpreterModule
 
@@ -56,26 +55,6 @@ def flatten_args(args):
     )
 
     return flat_args
-
-
-def _get_binary_filename(cur_idx: int, is_optim: bool = False) -> str:  # type: ignore[valid-type]
-    """
-    Gets filename for pytorch checkpoint binary based on current index and world size.
-
-    Args:
-        cur_idx (int): current device index
-        is_optim (bool): True if generating binary filename for optimizer,
-                         False otherwise
-
-    Returns:
-        str: checkpoint filename
-    """
-    idx = str(cur_idx + 1).zfill(5)
-    world_size = str(dist.get_world_size()).zfill(5)
-
-    state_type = "optim" if is_optim else "model"
-
-    return f"pytorch_{state_type}-{idx}-of-{world_size}.bin"
 
 
 def modify_graph_op_device(
